+++ pkg/db/ent/transaction.go
@@ -41,9 +41,9 @@ type Transaction struct {
 	// SignaturePlatform holds the value of the "signature_platform" field.
 	SignaturePlatform string `json:"signature_platform,omitempty"`
 	// CreatetimeUtc holds the value of the "createtime_utc" field.
-	CreatetimeUtc int `json:"createtime_utc,omitempty"`
+	CreatetimeUtc int64 `json:"createtime_utc,omitempty"`
 	// UpdatetimeUtc holds the value of the "updatetime_utc" field.
-	UpdatetimeUtc int `json:"updatetime_utc,omitempty"`
+	UpdatetimeUtc int64 `json:"updatetime_utc,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the TransactionQuery when eager-loading is set.
 	Edges                  TransactionEdges `json:"edges"`
@@ -196,13 +196,13 @@ func (t *Transaction) assignValues(columns []string, values []interface{}) error
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field createtime_utc", values[i])
 			} else if value.Valid {
-				t.CreatetimeUtc = int(value.Int64)
+				t.CreatetimeUtc = value.Int64
 			}
 		case transaction.FieldUpdatetimeUtc:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field updatetime_utc", values[i])
 			} else if value.Valid {
-				t.UpdatetimeUtc = int(value.Int64)
+				t.UpdatetimeUtc = value.Int64
 			}
 		case transaction.ForeignKeys[0]:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
